refactor: modularize executor library
